requirements.txt
emoji==1.6.1
numpy==1.21.2
pandas==1.3.3
PyYAML==5.4.1
scikit-learn
selenium==3.141.0
sentence-transformers==2.1.0